// src/constants/interface.ts
export interface Match {
  matchId: string;
  rowNo: string;
  typeName: string;
  matchTime: string;
  matchResult: string;
  recPercent: string;
  homeTeam: string;
  visitTeam: string;
  result1: string | null;
  result2: string | null;
  fixedNam: any;
  betRate: string;
  homeLogo: any;
  homeTeamNo: string | null;
  isOk: string;
  isCode: string | null;
  matchLong: string;
  visitLogo: any;
  visitTeamNo: any;
  week: any;
}